Events_management_api/requirements.txt
asgiref==3.10.0
certifi==2025.11.12
distlib==0.4.0
dj-database-url==3.1.0
Django==5.2.9
django-filter==25.2
django-taggit==6.1.0
djangorestframework==3.16.1
filelock==3.20.0
gunicorn==23.0.0
packaging==25.0
pipenv==2026.0.1
platformdirs==4.5.1
psycopg2-binary==2.9.11
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.35.4
whitenoise==6.11.0
